Fix scripts, examples, and configs that failed to run
- Fix broken imports: prepare_credit_card_data module path, nonexistent
  Trainer and calculate_metrics, and ReduceLROnPlateau(verbose=) removed in
  newer PyTorch.
- Add save_npz_data/load_npz_data with standard keys, alternate key names,
  and NaN padding for variable-length sequences; use them in safe-train,
  safe-eval, and the quick start script.
- safe-train sets model input_dim from the data and accepts an output path
  without a directory.
- Rewrite credit_card_config.yaml to match Config; Config now reports unknown
  keys and supports device 'auto', used by all bundled configs.
- Save the early-detection summary as plain floats so checkpoints load with
  torch.load(weights_only=True).
- Fix config paths in README and document the npz format.

+++ safe_fraud_detection/utils/metrics.py
@@ -167,9 +167,11 @@ class EarlyDetectionMetrics:
         if stats['num_fraudsters'] > 0:
             summary['early_detection_rate'] = stats['num_early_detected'] / stats['num_fraudsters']
         
+        # Plain Python floats so the summary can be saved in a checkpoint and
+        # loaded with torch.load(weights_only=True), which rejects numpy scalars
         if len(stats['early_detection_times']) > 0:
-            summary['avg_early_timestamps'] = np.mean(stats['early_detection_times'])
-            summary['median_early_timestamps'] = np.median(stats['early_detection_times'])
+            summary['avg_early_timestamps'] = float(np.mean(stats['early_detection_times']))
+            summary['median_early_timestamps'] = float(np.median(stats['early_detection_times']))
         
         return summary
